thru part 4.21, token auth required for delete

+++ part4/controllers/blogrouter.js
@@ -60,6 +60,17 @@ blogRouter.post('/', async (request, response) => {
   })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  if(!(decodedToken.id)){
+    return response.status(401).json({error: "token missing or invalid"})
+  }
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  if(!(user._id.toString() === blog.user.toString())){
+    return response.status(401).json({error: "token invalid"})
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
